requirements.txt
pyyaml
discord.py
sqlalchemy
